backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
ibm_watsonx_ai
supabase
ibm-watson
ibm-cos-sdk
ffmpeg-python